review(issue-91): address PR #92 review feedback
- parseEnum / parseFixed: inherit enclosing namespace when JSON omits one
- parseLogical: require "precision" for decimal logical type
- generic union decoder: bounds-check in int64 to avoid wrap on 32-bit
- Field JSON: round-trip Default and Order; reject unknown order strings
- Encoder/Decoder/doc.go: update concurrency note for io.Writer/io.Reader

+++ generic/codec.go
@@ -515,7 +515,7 @@ func compileUnion(u avro.Union, ctx *compileCtx) (*planNode, error) {
 			if err != nil {
 				return nil, err
 			}
-			if i < 0 || int(i) >= len(branches) {
+			if i < 0 || i >= int64(len(branches)) {
 				return nil, IndexOutOfRangeError{Kind: "union", Index: i, Len: len(branches)}
 			}
 			val, err := branches[i].dec(r)
